refactor: Centralize eraser size settings
